+++ src/main/kotlin/com/knu/mockin/model/dto/kisrequest/oauth/KISApprovalRequestDto.kt
@@ -1,9 +1,9 @@
-package com.knu.mockin.model.dto.request
+package com.knu.mockin.model.dto.kisrequest.oauth
 
 import com.fasterxml.jackson.annotation.JsonProperty
 
-data class KISTokenRequestDto(
+data class KISApprovalRequestDto(
     @JsonProperty("grant_type") val grantType: String,
     @JsonProperty("appkey") val appKey: String,
-    @JsonProperty("appsecret") val appSecret: String
+    @JsonProperty("secretkey") val secretKey: String
 )
