+++ src/main/kotlin/com/knu/mockin/model/dto/kisrequest/oauth/KISTokenRequestDto.kt
@@ -1,9 +1,9 @@
-package com.knu.mockin.model.dto.request
+package com.knu.mockin.model.dto.kisrequest.oauth
 
 import com.fasterxml.jackson.annotation.JsonProperty
 
-data class KISApprovalRequestDto(
+data class KISTokenRequestDto(
     @JsonProperty("grant_type") val grantType: String,
     @JsonProperty("appkey") val appKey: String,
-    @JsonProperty("secretkey") val secretKey: String
+    @JsonProperty("appsecret") val appSecret: String
 )
