+++ src/test/kotlin/com/knu/mockin/controller/AccountControllerTest.kt
@@ -1,39 +1,26 @@
 package com.knu.mockin.controller
 
 import com.knu.mockin.logging.utils.LogUtil
-import com.knu.mockin.model.dto.request.KISApprovalRequestDto
+import com.knu.mockin.model.dto.kisrequest.oauth.KISApprovalRequestDto
 import com.knu.mockin.model.dto.response.ApprovalKeyResponseDto
 import com.knu.mockin.service.AccountService
 import com.ninjasquad.springmockk.MockkBean
-import io.kotest.assertions.print.print
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.extensions.spring.SpringExtension
-import io.kotest.mpp.log
 import io.mockk.every
 import org.springframework.beans.factory.annotation.Value
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs
-import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.http.MediaType.APPLICATION_JSON
 import org.springframework.restdocs.ManualRestDocumentation
 import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation
 import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document
 import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration
-import org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint
 import org.springframework.restdocs.payload.JsonFieldType
 import org.springframework.restdocs.payload.PayloadDocumentation
 import org.springframework.restdocs.snippet.Snippet
-import org.springframework.test.context.web.WebAppConfiguration
 import org.springframework.test.web.reactive.server.WebTestClient
-import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.MockMvcResultHandlersDsl
 import org.springframework.test.web.servlet.client.MockMvcWebTestClient
-import org.springframework.test.web.servlet.post
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.asyncDispatch
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.*
-import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder
-import org.springframework.test.web.servlet.setup.MockMvcBuilders
 import org.springframework.web.context.WebApplicationContext
 import reactor.core.publisher.Mono
